+++ src/Matrix.php
@@ -1686,7 +1686,7 @@ class Matrix implements Tensor
         $b = $this->subtractColumnVector($mean);
 
         return $b->matmul($b->transpose())
-            ->divideScalar($this->m);
+            ->divideScalar($this->n);
     }
 
     /**